Cleanup and Icon ingame fix
-- Note, even though I have done all of this, you can still go back to previous commits and see how the files were back then, so you are not *losing* anything

 - Moved everything back 1 folder, don't need 2 folders like it is now
 - Removed plugins_data/ folder, don't need it
 - Removed deppth/ folder, don't need it either
 - Removed Firefly-Randomaster13/ folder inside img/ - This could've been causing an issue
 - I would advise you to move config.lua and co. into a folder called src/ to make it easier - but that is up to you
 - Moved FireflyEffectData and TraitFirefly into a folder called Boons/ and updated path in main.lua -- this is just to make it formatted a bit better and compact

 - Updated manifest.json to be correct - matters cause this is what the mod will see as the GUID
