--- Randomaster13-Firefly/README.md
@@ -1,3 +1,0 @@
-Massive credit to zanncdwbl for teaching me practically everything I needed to know for this mod!
-
-Currently in development and not working properly.
