--- Randomaster13-Firefly/config.lua
@@ -1,3 +1,0 @@
-return {
-    Enabled = true,
-}
