--- Randomaster13-Firefly/deppth/deppth-runner.py
@@ -1,6 +1,0 @@
-"""Convenience wrapper for running deppth directly from source tree."""
-
-from deppth.cli import main
-
-if __name__ == '__main__':
-  main()
