--- Randomaster13-Firefly/deppth/deppth/__main__.py
@@ -1,4 +1,0 @@
-"""__main__: executed when deppth directory is called as a script."""
-
-from .deppth import main
-main()
