--- Randomaster13-Firefly/img/Firefly-Randomaster13/manifest/Firefly-Randomaster130.atlas.json
@@ -1,1 +1,0 @@
-{"name": "bin\\Win\\Atlases\\Firefly-Randomaster130", "version": 4, "subAtlases": [], "isReference": true, "referencedTextureName": "bin\\Win\\Atlases\\Firefly-Randomaster130"}
