--- Randomaster13-Firefly/manifest.json
@@ -1,16 +1,0 @@
-{
-    "namespace": "randomaster",
-    "name": "Randomaster13-Firefly",
-    "description": "Adds features related to the song Fireflies by OwlCity.",
-    "version_number": "1.0.0",
-    "dependencies": [
-        "SGG_Modding-ModUtil-3.1.2",
-        "Hell2Modding-Hell2Modding-1.0.34",
-        "SGG_Modding-Chalk-2.1.1",
-        "SGG_Modding-SJSON-1.0.0",
-        "SGG_Modding-ReLoad-1.0.2",
-        "SGG_Modding-ENVY-1.0.1"
-    ],
-    "website_url": "https://github.com/randomaster13/Firefly-Boons",
-    "FullName": "Firefly_Boons"
-}
