+++ config.lua
@@ -1,0 +1,3 @@
+return {
+    Enabled = true,
+}
